refactor: extract vaadin-grid-filter logic into a reusable mixin
